+++ services/calculator/src/__tests__/express.store.errors.spec.ts
@@ -10,7 +10,7 @@ const exapp = app(new ExpressApp()).with(Calculator);
 jest.spyOn(store(), "query").mockRejectedValue(new Error("store query error"));
 jest.spyOn(store(), "stats").mockRejectedValue(new Error("store stats error"));
 
-describe("express app", () => {
+describe("calculator express app with store errors", () => {
   beforeAll(async () => {
     const express = exapp.build();
     express.get("/query", () => {
@@ -23,21 +23,19 @@ describe("express app", () => {
     await dispose()();
   });
 
-  describe("errors", () => {
-    it("should throw internal error on stream", async () => {
-      await expect(http.get("/all")).rejects.toThrowError();
-    });
+  it("should throw internal error on stream", async () => {
+    await expect(http.get("/all")).rejects.toThrowError();
+  });
 
-    it("should throw internal error on aggregate", async () => {
-      await expect(http.get("/calculator/test")).rejects.toThrowError("500");
-    });
+  it("should throw internal error on aggregate", async () => {
+    await expect(http.get("/calculator/test")).rejects.toThrowError("500");
+  });
 
-    it("should throw internal error on stats", async () => {
-      await expect(http.get("/stats")).rejects.toThrowError("500");
-    });
+  it("should throw internal error on stats", async () => {
+    await expect(http.get("/stats")).rejects.toThrowError("500");
+  });
 
-    it("should throw validation error", async () => {
-      await expect(http.get("/query")).rejects.toThrowError("400");
-    });
+  it("should throw validation error", async () => {
+    await expect(http.get("/query")).rejects.toThrowError("400");
   });
 });
